WIP: Test Congress Doctor driver support
WIP 1: Congress rules need to be fixed
WIP 2: test both inspectors (sample and Congress) in single functest run
WIP 3: make sure Congress Nova datasource is created with api_version >=
2.11

JIRA: DOCTOR-56

+++ tests/run.sh
@@ -29,6 +29,9 @@ SUPPORTED_INSTALLER_TYPES="apex fuel local"
 INSTALLER_TYPE=${INSTALLER_TYPE:-local}
 INSTALLER_IP=${INSTALLER_IP:-none}
 
+SUPPORTED_INSPECTOR_TYPES="sample congress"
+INSPECTOR_TYPE=${INSPECTOR_TYPE:-sample}
+
 ssh_opts="-o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no"
 as_doctor_user="--os-username $DOCTOR_USER --os-password $DOCTOR_PW
                 --os-tenant-name $DOCTOR_PROJECT"
@@ -182,7 +185,7 @@ print_log() {
 
 start_monitor() {
     pgrep -f "python monitor.py" && return 0
-    sudo python monitor.py "$COMPUTE_HOST" "$COMPUTE_IP" \
+    sudo -E python monitor.py "$COMPUTE_HOST" "$COMPUTE_IP" "$INSPECTOR_TYPE" \
         "http://127.0.0.1:$INSPECTOR_PORT/events" > monitor.log 2>&1 &
 }
 
@@ -192,15 +195,61 @@ stop_monitor() {
     print_log monitor.log
 }
 
+congress_add_rule() {
+    name=$1
+    policy=$2
+    rule=$3
+
+    if ! openstack congress policy rule list $policy | grep -q -e "// Name: $name$" ; then
+        openstack congress policy rule create --name $name $policy "$rule"
+    fi
+}
+
+congress_del_rule() {
+    name=$1
+    policy=$2
+
+    if openstack congress policy rule list $policy | grep -q -e "^// Name: $name$" ; then
+        openstack congress policy rule delete $policy $name
+    fi
+}
+
+congress_setup_rules() {
+    congress_add_rule nic_down classification \
+        'nic_down(host) :-
+            doctor:events(hostname=host, type="host.nic1.down")'
+
+    congress_add_rule host_force_down classification \
+        'execute[nova:services.force_down(host, "nova-compute")] :-
+            nic_down(host)'
+
+    congress_add_rule error_vms_states classification \
+        'execute[nova:servers.reset_state(vm)] :-
+            nova:servers(id=vm, host_name=host),
+            nic_down(host)'
+            #nova:services(host=x, binary="nova-compute", state="down")'
+}
+
 start_inspector() {
-    pgrep -f "python inspector.py" && return 0
-    python inspector.py "$INSPECTOR_PORT" > inspector.log 2>&1 &
+    if [[ "$INSPECTOR_TYPE" == "sample" ]] ; then
+        pgrep -f "python inspector.py" && return 0
+        python inspector.py "$INSPECTOR_PORT" > inspector.log 2>&1 &
+    elif [[ "$INSPECTOR_TYPE" == "congress" ]] ; then
+        openstack congress driver list | grep -q " doctor "
+        openstack congress datasource list | grep -q " doctor " || {
+            openstack congress datasource create doctor doctor
+        }
+        # TODO(cgoncalves): nova DS api_version=2.11
+        congress_setup_rules
+    fi
 }
 
 stop_inspector() {
-    pgrep -f "python inspector.py" || return 0
-    kill $(pgrep -f "python inspector.py")
-    print_log inspector.log
+    if [[ "$INSPECTOR_TYPE" == "sample" ]] ; then
+        pgrep -f "python inspector.py" || return 0
+        kill $(pgrep -f "python inspector.py")
+        print_log inspector.log
+    fi
 }
 
 start_consumer() {
@@ -303,6 +352,10 @@ cleanup() {
                               --project "$DOCTOR_PROJECT"
     openstack project delete "$DOCTOR_PROJECT"
     openstack user delete "$DOCTOR_USER"
+
+    congress_del_rule error_vms_states classification
+    congress_del_rule host_force_down classification
+    congress_del_rule nic_down classification
 }
